feat: finish module
