+++ src/modules/esm.mjs
@@ -1,0 +1,51 @@
+import path from 'path';
+import fs from 'node:fs/promises';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+import './files/c.js';
+import { fileURLToPath } from 'url';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+const  getDataFromJson = async (path) => {
+    const data = await fs.readFile(path, 'utf-8');
+    return JSON.parse(data);
+};
+
+const firstJson = await getDataFromJson(path.join(__dirname, 'files/a.json'));
+const secondJson = await getDataFromJson(path.join(__dirname, 'files/b.json'));
+
+
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+    unknownObject = firstJson;
+} else {
+    unknownObject = secondJson;
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log('To terminate it, use Ctrl+C combination');
+});
+
+export { unknownObject, myServer};
+
